+++ DTCD-StyleSystem/src/base-components/BaseColorPicker/BaseColorPicker.js
@@ -175,16 +175,12 @@ export default class BaseColorPicker extends HTMLElement {
     }
   }
 
-  #handleDocumentClick = event => {
-    let isPickerContainsOriginalTarget = false;
-    try {
-      isPickerContainsOriginalTarget = this.#picker.contains(event.originalTarget);
-    } catch (error) {}
-
-    let isComponentContainsTarget = this.contains(event.target);
-    const resultCondition = !isPickerContainsOriginalTarget && !isComponentContainsTarget;
-
-    if (resultCondition) {
+  #handleDocumentClick = (event) => {
+    const eventPath = event.composedPath();
+    const originalTarget = eventPath[0];
+    const isPickerContainsOriginalTarget = this.#picker.contains(originalTarget);
+    const isComponentContainsTarget = this.contains(event.target);
+    if (!isPickerContainsOriginalTarget && !isComponentContainsTarget) {
       this.toggle(false);
     }
   };