+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - BaseSelect option list correct opening
+- BaseColorPicker outside click in Chrome
 
 ## [0.5.0]